+++ ochre-api/src/main/java/gov/vha/isaac/ochre/api/DatabaseServices.java
@@ -1,6 +1,5 @@
 package gov.vha.isaac.ochre.api;
 
-import java.nio.file.Files;
 import java.nio.file.Path;
 
 import org.jvnet.hk2.annotations.Contract;