+++ ochre-api/src/main/java/gov/vha/isaac/ochre/api/externalizable/MultipleDataWriterService.java
@@ -20,8 +20,12 @@ package gov.vha.isaac.ochre.api.externalizable;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 import org.apache.logging.log4j.LogManager;
 import org.slf4j.Logger;
@@ -29,7 +33,11 @@ import org.slf4j.LoggerFactory;
 import gov.vha.isaac.ochre.api.LookupService;
 
 /**
- * Simple wrapper class to allow us to serialize to multiple formats at once
+ * Simple wrapper class to allow us to serialize to multiple formats at once.
+ * 
+ * Also includes logic for incorporating the date and a UUID into the file name to ensure uniqueueness, 
+ * and logic for rotating the changeset files.
+ * 
  * {@link MultipleDataWriterService}
  *
  * @author <a href="mailto:daniel.armbrust.list@gmail.com">Dan Armbrust</a>
@@ -38,9 +46,67 @@ public class MultipleDataWriterService implements DataWriterService
 {
 	ArrayList<DataWriterService> writers_ = new ArrayList<>();
 	private Logger logger = LoggerFactory.getLogger(MultipleDataWriterService.class);
+	private final SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd_HH-mm-ss");
+	private String prefix;
+	private AtomicInteger objectWriteCount = new AtomicInteger();
+	private final int rotateAfter = 10000;  //This will cause us to rotate files after ~ 1 MB of IBDF content, in rough testing.
+	private boolean enableRotate;
 	
+	/**
+	 * This constructor sets up the multipleDataWriter in such a way that is will create date stamped and UUID unique file names, rotating them after 
+	 * a certain number of writes, to prevent them from growing too large.
+	 * 
+	 * This constructor will also start a mode where we do NOT keep 0 length files - therefore, if we start, and stop, and the last file that was being written
+	 * to is size 0, the last file will be deleted.
+	 * 
+	 * @param folderToWriteInto
+	 * @param prefix
+	 * @param jsonExtension
+	 * @param ibdfExtension
+	 * @throws IOException
+	 */
+	public MultipleDataWriterService(Path folderToWriteInto, String prefix, Optional<String> jsonExtension, Optional<String> ibdfExtension) throws IOException
+	{
+		this.prefix = prefix;
+		enableRotate = true;
+		String fileNamePrefix = prefix + sdf.format(new Date()) + "_" + UUID.randomUUID().toString() + ".";
+		if (jsonExtension.isPresent())
+		{
+			//Use HK2 here to make fortify stop false-flagging an open resource error
+			DataWriterService writer = LookupService.get().getService(DataWriterService.class, "jsonWriter");
+			if (writer != null)
+			{
+				writer.configure(folderToWriteInto.resolve(fileNamePrefix + jsonExtension.get()));
+				writers_.add(writer);
+			}
+			else
+			{
+				LogManager.getLogger().warn("json writer was requested, but not found on classpath!");
+			}
+		}
+		if (ibdfExtension.isPresent())
+		{
+			DataWriterService writer = LookupService.get().getService(DataWriterService.class, "ibdfWriter");
+			if (writer != null)
+			{
+				writer.configure(folderToWriteInto.resolve(fileNamePrefix + ibdfExtension.get()));
+				writers_.add(writer);
+			}
+			else
+			{
+				LogManager.getLogger().warn("ibdf writer was requested, but not found on classpath!");
+			}
+		}
+	}
+	
+	/**
+	 * This constructor creates a multiple data writer service which writes to the specified files, and does not do any rotation or autonaming.
+	 * @param jsonPath
+	 * @param ibdfPath
+	 */
 	public MultipleDataWriterService(Optional<Path> jsonPath, Optional<Path> ibdfPath) throws IOException
 	{
+		enableRotate = false;
 		if (jsonPath.isPresent())
 		{
 			//Use HK2 here to make fortify stop false-flagging an open resource error
@@ -55,6 +121,7 @@ public class MultipleDataWriterService implements DataWriterService
 				LogManager.getLogger().warn("json writer was requested, but not found on classpath!");
 			}
 		}
+		
 		if (ibdfPath.isPresent())
 		{
 			DataWriterService writer = LookupService.get().getService(DataWriterService.class, "ibdfWriter");
@@ -67,6 +134,28 @@ public class MultipleDataWriterService implements DataWriterService
 			{
 				LogManager.getLogger().warn("ibdf writer was requested, but not found on classpath!");
 			}
+		}
+	}
+
+	private void rotateFiles() throws RuntimeException
+	{
+		try
+		{
+			pause();
+			String fileNamePrefix = prefix + sdf.format(new Date()) + "_" + UUID.randomUUID().toString();
+			for (DataWriterService writer : writers_)
+			{
+				String extension = writer.getCurrentPath().getFileName().toString();
+				extension = extension.substring(extension.lastIndexOf('.'));
+				writer.configure(writer.getCurrentPath().getParent().resolve(fileNamePrefix + extension));
+			}
+			objectWriteCount.set(0);
+			resume();
+		}
+		catch (IOException e)
+		{
+			logger.error("Unexpected error rotating changeset files!", e);
+			throw new RuntimeException(e);
 		}
 	}
 	
@@ -103,6 +192,10 @@ public class MultipleDataWriterService implements DataWriterService
 				logger.warn("Unexpected", e);
 				throw new RuntimeException(e);
 			}
+		}
+		if (enableRotate && objectWriteCount.incrementAndGet() >= rotateAfter)
+		{
+			rotateFiles();
 		}
 	}
 	
@@ -218,6 +311,12 @@ public class MultipleDataWriterService implements DataWriterService
 	 */
 	@Override
 	public void configure(Path path) throws UnsupportedOperationException
+	{
+		throw new UnsupportedOperationException("Method not supported");
+	}
+
+	@Override
+	public Path getCurrentPath()
 	{
 		throw new UnsupportedOperationException("Method not supported");
 	}