+++ sync-git/src/main/java/gov/va/isaac/sync/git/gitblit/GitBlitUtils.java
@@ -20,10 +20,14 @@ package gov.va.isaac.sync.git.gitblit;
 
 import java.io.IOException;
 import java.util.Date;
+import java.util.HashSet;
+import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import gov.va.isaac.sync.git.gitblit.models.RepositoryModel;
 import gov.va.isaac.sync.git.gitblit.utils.RpcUtils;
+import gov.va.isaac.sync.git.gitblit.utils.RpcUtils.AccessRestrictionType;
+
 
 /**
  * {@link GitBlitUtils}
@@ -51,18 +55,58 @@ public class GitBlitUtils
 	 * @return true if the action succeeded
 	 * @throws IOException
 	 */	
-	public static boolean createRepository(String baseRemoteAddress, String repoName, String repoDesc, String username, char[] password) throws IOException
+	public static void createRepository(String baseRemoteAddress, String repoName, String repoDesc, String username, char[] password, boolean allowRead) throws IOException
 	{
 		try
 		{
-			boolean status =  RpcUtils.createRepository(new RepositoryModel(repoName, repoDesc, username, new Date()), baseRemoteAddress, username, password);
+			RepositoryModel rm = new RepositoryModel(repoName, repoDesc, username, new Date());
+			if (allowRead)
+			{
+				rm.accessRestriction = AccessRestrictionType.PUSH.toString();
+			}
+			
+			boolean status =  RpcUtils.createRepository(rm, adjustUrlForGitBlit(baseRemoteAddress),
+					username, password);
 			log.info("Repository: "+repoName +", create successfully: " + status);
-			return status;
+			if (!status)
+			{
+				throw new IOException("Create of repo '" + repoName + "' failed");
+			}
 		}
 		catch (Exception e)
 		{
 			log.error("Failed to create repository: "+repoName +", Unexpected Error: ", e);
-			throw new IOException("Failed to create repository: "+repoName +",Internal error", e);
+			throw new IOException("Failed to create repository: "+repoName +", Internal error", e);
 		}
+	}
+	
+	public static Set<String> readRepositories(String baseRemoteAddress, String username, char[] password) throws IOException
+	{
+		HashSet<String> results = new HashSet<>();
+		RpcUtils.getRepositories(adjustUrlForGitBlit(baseRemoteAddress), username, password).forEach((name, value) -> results.add((String)value.get("name")));
+		return results;
+	}
+	
+	/**
+	 * This hackery is being done because of a code-sync issue between PRISME and ISAAC-Rest, where PRISME is putting a bare URL into the props file.
+	 * It will be fixed on the PRISME side, eventually, making this method a noop - but for now, handle either the old or new style.
+	 * 
+	 * Essentially, if we see a bare URL like https://vaauscttdbs80.aac.va.gov:8080 we add /git to the end of it.
+	 * If we see a URL that includes a location - like https://vaauscttdbs80.aac.va.gov:8080/gitServer - we do nothing more than add a trailing forward slash
+	 * @param url
+	 * @return
+	 */
+	public static String adjustUrlForGitBlit(String url)
+	{
+		String temp = url;
+		if (!temp.endsWith("/"))
+		{
+			temp += "/";
+		}
+		if (temp.matches("https?:\\/\\/[a-zA-Z0-9\\.]+:?\\d*\\/$"))
+		{
+			temp += "git/";
+		}
+		return temp;
 	}
 }