+++ sync-git/src/main/java/gov/va/isaac/sync/git/gitblit/utils/JsonUtils.java
@@ -1,36 +1,25 @@
 package gov.va.isaac.sync.git.gitblit.utils;
 
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.OutputStream;
-import java.lang.reflect.Type;
 import java.net.HttpURLConnection;
 import java.net.URLConnection;
 import java.text.DateFormat;
-import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.Date;
+import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 import java.util.TimeZone;
-
-import com.google.gson.ExclusionStrategy;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonPrimitive;
-import com.google.gson.JsonSerializationContext;
-import com.google.gson.JsonSerializer;
-import com.google.gson.JsonSyntaxException;
-import com.google.gson.reflect.TypeToken;
-
+import com.cedarsoftware.util.io.JsonObject;
+import com.cedarsoftware.util.io.JsonReader;
+import com.cedarsoftware.util.io.JsonWriter;
 import gov.va.isaac.sync.git.gitblit.GitBlitException.ForbiddenException;
 import gov.va.isaac.sync.git.gitblit.GitBlitException.NotAllowedException;
 import gov.va.isaac.sync.git.gitblit.GitBlitException.UnauthorizedException;
 import gov.va.isaac.sync.git.gitblit.GitBlitException.UnknownRequestException;
-import gov.va.isaac.sync.git.gitblit.models.RepositoryModel;
-import gov.va.isaac.sync.git.gitblit.utils.RpcUtils.AccessPermission;
 
 /**
  * Utility methods for json calls to a Gitblit server.
@@ -38,10 +27,8 @@ import gov.va.isaac.sync.git.gitblit.utils.RpcUtils.AccessPermission;
  * @author James Moger
  *
  */
-public class JsonUtils {
-
-	public static final Type REPOSITORIES_TYPE = new TypeToken<Map<String, RepositoryModel>>() {
-	}.getType();
+public class JsonUtils
+{
 
 	/**
 	 * Creates JSON from the specified object.
@@ -49,26 +36,34 @@ public class JsonUtils {
 	 * @param o
 	 * @return json
 	 */
-	public static String toJsonString(Object o) {
-		String json = gson().toJson(o);
-		return json;
+	public static String toJsonString(Object o)
+	{
+		DateFormat dateformat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'", Locale.US);
+		dateformat.setTimeZone(TimeZone.getTimeZone("UTC"));
+		HashMap<String, Object> config = new HashMap<>();
+		config.put(JsonWriter.DATE_FORMAT, dateformat);
+		config.put(JsonWriter.SKIP_NULL_FIELDS, true);
+		config.put(JsonWriter.TYPE, false);
+		config.put(JsonWriter.ENUM_PUBLIC_ONLY, true);
+		return JsonWriter.objectToJson(o, config);
 	}
 
 	/**
 	 * Sends a JSON message.
 	 *
 	 * @param url
-	 *            the url to write to
+	 * the url to write to
 	 * @param json
-	 *            the json message to send
+	 * the json message to send
 	 * @param username
 	 * @param password
 	 * @return the http request result code
 	 * @throws {@link IOException}
 	 */
-	public static int sendJsonString(String url, String json, String username, char[] password)
-			throws IOException {
-		try {
+	public static int sendJsonString(String url, String json, String username, char[] password) throws IOException
+	{
+		try
+		{
 			byte[] jsonBytes = json.getBytes(ConnectionUtils.CHARSET);
 			URLConnection conn = ConnectionUtils.openConnection(url, username, password);
 			conn.setRequestProperty("Content-Type", "text/plain;charset=" + ConnectionUtils.CHARSET);
@@ -82,17 +77,26 @@ public class JsonUtils {
 
 			int status = ((HttpURLConnection) conn).getResponseCode();
 			return status;
-		} catch (IOException e) {
-			if (e.getMessage().indexOf("401") > -1) {
+		}
+		catch (IOException e)
+		{
+			if (e.getMessage().indexOf("401") > -1)
+			{
 				// unauthorized
 				throw new UnauthorizedException(url);
-			} else if (e.getMessage().indexOf("403") > -1) {
+			}
+			else if (e.getMessage().indexOf("403") > -1)
+			{
 				// requested url is forbidden by the requesting user
 				throw new ForbiddenException(url);
-			} else if (e.getMessage().indexOf("405") > -1) {
+			}
+			else if (e.getMessage().indexOf("405") > -1)
+			{
 				// requested url is not allowed by the server
 				throw new NotAllowedException(url);
-			} else if (e.getMessage().indexOf("501") > -1) {
+			}
+			else if (e.getMessage().indexOf("501") > -1)
+			{
 				// requested url is not recognized by the server
 				throw new UnknownRequestException(url);
 			}
@@ -100,65 +104,78 @@ public class JsonUtils {
 		}
 	}
 
-	// build custom gson instance with GMT date serializer/deserializer
-	// http://code.google.com/p/google-gson/issues/detail?id=281
-	public static Gson gson(ExclusionStrategy... strategies) {
-		GsonBuilder builder = new GsonBuilder();
-		builder.registerTypeAdapter(Date.class, new GmtDateTypeAdapter());
-		builder.registerTypeAdapter(AccessPermission.class, new AccessPermissionTypeAdapter());
-		if (strategies != null && strategies.length > 0) {
-			builder.setExclusionStrategies(strategies);
+	/**
+	 * Reads a gson object from the specified url.
+	 *
+	 * @param url
+	 * @param type
+	 * @param username
+	 * @param password
+	 * @return the deserialized object
+	 * @throws {@link IOException}
+	 */
+	@SuppressWarnings("unchecked")
+	public static JsonObject<String, Map<String, ?>> retrieveJson(String url, String username, char[] password) throws IOException
+	{
+		String json = retrieveJsonString(url, username, password);
+		if (StringUtils.isEmpty(json))
+		{
+			return null;
 		}
-		return builder.create();
+		DateFormat dateformat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'", Locale.US);
+		dateformat.setTimeZone(TimeZone.getTimeZone("UTC"));
+		HashMap<String, Object> config = new HashMap<>();
+		config.put(JsonWriter.DATE_FORMAT, dateformat);
+		return (JsonObject<String, Map<String, ?>>) JsonReader.jsonToJava(json, config);
 	}
 
-	private static class AccessPermissionTypeAdapter implements JsonSerializer<AccessPermission>, JsonDeserializer<AccessPermission> {
-
-		private AccessPermissionTypeAdapter() {
-		}
-
-		@Override
-		public synchronized JsonElement serialize(AccessPermission permission, Type type,
-				JsonSerializationContext jsonSerializationContext) {
-			return new JsonPrimitive(permission.code);
-		}
-
-		@Override
-		public synchronized AccessPermission deserialize(JsonElement jsonElement, Type type,
-				JsonDeserializationContext jsonDeserializationContext) {
-			return AccessPermission.fromCode(jsonElement.getAsString());
-		}
-	}
-	public static class GmtDateTypeAdapter implements JsonSerializer<Date>, JsonDeserializer<Date> {
-		private final DateFormat dateFormat;
-
-		public GmtDateTypeAdapter() {
-			dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'", Locale.US);
-			dateFormat.setTimeZone(TimeZone.getTimeZone("UTC"));
-		}
-
-		@Override
-		public synchronized JsonElement serialize(Date date, Type type,
-				JsonSerializationContext jsonSerializationContext) {
-			synchronized (dateFormat) {
-				String dateFormatAsString = dateFormat.format(date);
-				return new JsonPrimitive(dateFormatAsString);
+	/**
+	 * Retrieves a JSON message.
+	 *
+	 * @param url
+	 * @return the JSON message as a string
+	 * @throws {@link IOException}
+	 */
+	public static String retrieveJsonString(String url, String username, char[] password) throws IOException
+	{
+		try
+		{
+			URLConnection conn = ConnectionUtils.openReadConnection(url, username, password);
+			InputStream is = conn.getInputStream();
+			BufferedReader reader = new BufferedReader(new InputStreamReader(is, ConnectionUtils.CHARSET));
+			StringBuilder json = new StringBuilder();
+			char[] buffer = new char[4096];
+			int len = 0;
+			while ((len = reader.read(buffer)) > -1)
+			{
+				json.append(buffer, 0, len);
 			}
+			is.close();
+			return json.toString();
 		}
-
-		@Override
-		public synchronized Date deserialize(JsonElement jsonElement, Type type,
-				JsonDeserializationContext jsonDeserializationContext) {
-			try {
-				synchronized (dateFormat) {
-					Date date = dateFormat.parse(jsonElement.getAsString());
-					return new Date((date.getTime() / 1000) * 1000);
-				}
-			} catch (ParseException e) {
-				throw new JsonSyntaxException(jsonElement.getAsString(), e);
+		catch (IOException e)
+		{
+			if (e.getMessage().indexOf("401") > -1)
+			{
+				// unauthorized
+				throw new UnauthorizedException(url);
 			}
+			else if (e.getMessage().indexOf("403") > -1)
+			{
+				// requested url is forbidden by the requesting user
+				throw new ForbiddenException(url);
+			}
+			else if (e.getMessage().indexOf("405") > -1)
+			{
+				// requested url is not allowed by the server
+				throw new NotAllowedException(url);
+			}
+			else if (e.getMessage().indexOf("501") > -1)
+			{
+				// requested url is not recognized by the server
+				throw new UnknownRequestException(url);
+			}
+			throw e;
 		}
 	}
-
-
 }